// packages/platform/src/types/mt5/account/user.type.ts
export interface Mt5User {
  login: number;
  group: string;
  certSerialNumber: number;
  rights: string;
  registration: number;
  lastAccess: number;
  lastIP: string;
  name: string;
  company: string;
  account: string;
  country: string;
  language: number;
  city: string;
  state: string;
  zipCode: string;
  address: string;
  phone: string;
  eMail: string;
  id: string;
  status: string;
  comment: string;
  color: number;
  phonePassword: string;
  leverage: number;
  agent: number;
  balance: number;
  credit: number;
  interestRate: number;
  commissionDaily: number;
  commissionMonthly: number;
  commissionAgentDaily: number;
  commissionAgentMonthly: number;
  balancePrevDay: number;
  balancePrevMonth: number;
  equityPrevDay: number;
  equityPrevMonth: number;
  lastPassChange: number;
  leadCampaign: string;
  leadSource: string;
  externalAccountTotal: number;
  mqid: string;
  clientID: number;
  firstName: string;
  lastName: string;
  middleName: string;
}